+++ src/js/backend.ts
@@ -38,6 +38,12 @@ export interface ReactNativeOptions extends BrowserOptions {
 
   /** The interval to end a session if the App goes to the background. */
   sessionTrackingIntervalMillis?: number;
+
+  /** Enable scope sync from Java to NDK on Android */
+  enableNdkScopeSync?: boolean;
+
+  /** When enabled, all the threads are automatically attached to all logged events on Android */
+  attachThreads?: boolean;
 }
 
 /** The Sentry ReactNative SDK Backend. */